backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pandas==2.1.1
pyarrow==14.0.1
duckdb==0.9.2
redis==5.0.1
python-dotenv==1.0.0
pydantic==2.7.4
gunicorn==21.2.0
httpx==0.28.1
pytest==7.4.3
numpy==1.25.2
requests==2.31.0
python-dotenv==1.0.0
upstash-redis